+++ src/ServicePulse.Core/ServicePulseExtensions.cs
@@ -1,6 +1,7 @@
 ﻿namespace ServicePulse;
 
 using System.Net.Mime;
+using System.Reflection;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
@@ -27,9 +28,9 @@ public static class ServicePulseExtensions
             });
     }
 
-    public static IEndpointRouteBuilder MapServicePulseConstants(this IEndpointRouteBuilder app, ServicePulseSettings constants)
+    public static IEndpointRouteBuilder MapServicePulseConstants(this IEndpointRouteBuilder app, ServicePulseSettings settings)
     {
-        var constantsFile = constants.GetConstantsFileContents();
+        var constantsFile = GetConstantsFileContents(settings);
 
         app.MapGet("/js/app.constants.js", (HttpContext context) =>
         {
@@ -38,5 +39,32 @@ public static class ServicePulseExtensions
         });
 
         return app;
+    }
+
+    static string GetConstantsFileContents(ServicePulseSettings settings)
+    => $$"""
+        window.defaultConfig = {
+            default_route: '{{settings.DefaultRoute}}',
+            version: '{{GetVersionInformation()}}',
+            service_control_url: '{{settings.ServiceControlUrl}}',
+            monitoring_urls: ['{{settings.MonitoringUrl ?? "!"}}'],
+            showPendingRetry: {{(settings.ShowPendingRetry ? "true" : "false")}},
+        }
+        """;
+    static string GetVersionInformation()
+    {
+        var majorMinorPatch = "0.0.0";
+
+        var attributes = typeof(ServicePulseExtensions).Assembly.GetCustomAttributes<AssemblyMetadataAttribute>();
+
+        foreach (var attribute in attributes)
+        {
+            if (attribute.Key == "MajorMinorPatch")
+            {
+                majorMinorPatch = attribute.Value ?? "0.0.0";
+            }
+        }
+
+        return majorMinorPatch;
     }
 }
